requirements.txt
flask
flask-login
flask-bcrypt
python-dotenv
google-genai
pandas
openpyxl